api/requirements.txt
httptools
uvloop
shippedbrain
aiohttp==3.6.3
bcrypt==3.2.0
click==7.1.2
cryptography==3.1.1
docker==4.3.1
ecdsa==0.14.1
entrypoints==0.3
fastapi==0.61.1
fastapi-utils
Flask==1.1.2
gunicorn==20.0.4
Jinja2==2.11.2
mlflow==1.14.1
msrest==0.6.19
multidict==4.7.6
numpy==1.19.2
oauthlib==3.1.0
pandas==1.2
path==15.0.0
protobuf==3.13.0
psycopg2-binary==2.8.6
pycparser==2.20
pydantic==1.6.1
pylint==2.6.0
python-dateutil==2.8.1
python-dotenv==0.14.0
python-editor==1.0.4
python-jose==3.2.0
python-multipart==0.0.5
pytz==2020.1
PyYAML==5.3.1
requests==2.24.0
requests-oauthlib==1.3.0
SQLAlchemy==1.3.19
urllib3==1.25.10
uvicorn==0.12.1
boto3
